Random moved to a random component, shooting logic is inshooting clas…

+++ A19 Ex01 HeziDebby 203796701 GalNahum 312535644/GameObject.cs
@@ -16,7 +16,7 @@ namespace A19_Ex01_HeziDebby_203796701_GalNahum_312535644
         protected ContentManager Content { get; private set; }
 
         public Vector2 CurrentPosition { get; set; }
-        public int Velocity { get; set; }
+        public float Velocity { get; set; }
         public Texture2D Texture { get; protected set; }
         public Utilities.eDirection CurrentDirection { get; set; }
         public Color Color { get; set; }
